+++ packages/cli/src/generate/command.ts
@@ -3,54 +3,101 @@ import {
   DeployedContract,
   fetchContractMetadataFromAddress,
   getChainProvider,
-  ThirdwebSDK,
 } from "@thirdweb-dev/sdk";
 import { ThirdwebStorage } from "@thirdweb-dev/storage";
+import { spinner, info } from "../core/helpers/logger";
+import { allChains } from "@thirdweb-dev/chains";
 import fs from "fs";
 import prompts from "prompts";
-
-type GenerateOptions = {
-  path: string;
-  deployer?: string;
-};
+import { findMatches } from "../common/file-helper";
+import ora from "ora";
+import { GenerateOptions, ThirdwebConfig } from "./types";
+import { CHAIN_OPTIONS, getContractsForAddresses } from "./utils";
 
 export async function generate(options: GenerateOptions) {
   let projectPath: string = options.path?.replace(/\/$/, "") || ".";
-
   let contracts: DeployedContract[] = [];
+
+  // Find all addresses in this project
+  let addresses: string[] = [];
+  findMatches(projectPath, /(0x[a-fA-F0-9]{40})/g, addresses);
+
+  // We check if there's a thirdweb.json config file present
   if (fs.existsSync(`${projectPath}/thirdweb.json`)) {
-    // First we check if there's a thirdweb.json config file present
-    const thirdwebConfig = JSON.parse(
+    // Read the current thirdweb config file
+    const thirdwebConfig: ThirdwebConfig = JSON.parse(
       fs.readFileSync(`${projectPath}/thirdweb.json`, "utf-8"),
     );
-    contracts = thirdwebConfig.contracts as DeployedContract[];
-  } else {
-    // Otherwise, we get contracts by deployer address (and generate a thirdweb.json)
-    let deployerAddress: string;
-    if (options.deployer) {
-      deployerAddress = options.deployer;
-    } else {
-      const res = await prompts({
-        type: "text",
-        name: "deployer",
-        message: GENERATE_MESSAGES.deployerAddress,
-      });
 
-      deployerAddress = res.deployer.trim();
-    }
-
-    const sdk = new ThirdwebSDK("polygon");
-    contracts = await sdk.multiChainRegistry.getContractAddresses(
-      deployerAddress,
+    // Check if there are any addresses in the project that aren't yet in the config
+    const configAddresses = thirdwebConfig.contracts.map((contract) =>
+      contract.address.toLowerCase(),
+    );
+    const newAddresses = addresses.filter(
+      (address) => !configAddresses.includes(address.toLowerCase()),
     );
 
+    // Initialize contracts to the contracts in thirdweb.json file
+    contracts = [...thirdwebConfig.contracts];
+
+    // If we have new contracts that aren't yet stored in config
+    // get full contract objects
+    if (newAddresses.length) {
+      const chainIds = thirdwebConfig.chainIds;
+      await getContractsForAddresses(newAddresses, chainIds, contracts);
+
+      if (contracts.length > thirdwebConfig.contracts.length) {
+        fs.writeFileSync(
+          `${projectPath}/thirdweb.json`,
+          JSON.stringify({ chainIds, contracts }, undefined, 2),
+        );
+
+        const numberOfNewContracts =
+          contracts.length - thirdwebConfig.contracts.length;
+        info(
+          `Updated thirdweb.json with ${numberOfNewContracts} new contract${
+            numberOfNewContracts === 1 ? "" : "s"
+          }`,
+        );
+      }
+    }
+  } else {
+    // If there is no configuration file present, we need to generate one...
+
+    // First we ask the user what chains are used in their project
+    const res = await prompts({
+      type: "autocompleteMultiselect",
+      name: "chains",
+      message: GENERATE_MESSAGES.chains,
+      choices: CHAIN_OPTIONS,
+    });
+
+    const chainIds: number[] = res.chains.map(
+      (index: number) => allChains[index].chainId,
+    );
+
+    // Then we check for the chainId of each address
+    await getContractsForAddresses(addresses, chainIds, contracts);
+
+    // Create a thirdweb.json file with the specific contracts and chain ids
     fs.writeFileSync(
       `${projectPath}/thirdweb.json`,
-      JSON.stringify({ contracts }, undefined, 2),
+      JSON.stringify({ chainIds, contracts }, undefined, 2),
+    );
+
+    info(
+      `Created a thirdweb.json file with configuration for ${
+        contracts.length
+      } contract${
+        contracts.length === 1 ? "" : "s"
+      } detected in your project.\n\n - You can also update this configuration manually by editing the file.\n`,
     );
   }
 
-  // Attempt to download the ABI for the contract
+  // Attempt to download the ABI for each contract
+  const abiSpinner = spinner(
+    `Downloading ABIs for contracts configured in 'thirdweb.json'`,
+  );
   const storage = new ThirdwebStorage();
   const metadata: {
     address: string;
@@ -129,6 +176,12 @@ export async function generate(options: GenerateOptions) {
     fs.writeFileSync(typeFilePath, updatedFile);
   }
 
+  abiSpinner.succeed(
+    `Downloaded and cached ABIs for ${contracts.length} smart contract${
+      contracts.length === 1 ? "" : "s"
+    }`,
+  );
+
   // Add generate command to postinstall
   const packageJsonPath = `${projectPath}/package.json`;
   if (!fs.existsSync(packageJsonPath)) {
@@ -159,4 +212,8 @@ export async function generate(options: GenerateOptions) {
       2,
     ),
   );
+
+  ora(
+    "Added 'npx thirdweb generate' to postinstall in package.json.\n\n - This is necessary to use 'thirdweb generate' in production.\n",
+  ).info();
 }